Remove test duplication related to TestCase inheritance in different files
